Started refactoring index.js
