+++ src/loggingUtil.js
@@ -1,0 +1,6 @@
+import * as log from 'loglevel'
+
+const logger = log.getLogger('disciplLawReg')
+export function getDiscplLogger () {
+  return logger
+}
